requirements.txt
yt-dlp
openai
openai-whisper
pydub
librosa
opencv-python
pytesseract
numpy
Pillow
pymupdf
flask
flask_session
cs50
transformers
torch
flask-wtf
flask-login
flask-SQLAlchemy
python-dotenv
flask-mail
flask-migrate
flask-cors
flask-restful
flask-security
flask-bcrypt
flask-principal
Flask-JWT-Extended
celery
redis
pytest
pytest-flask
flask-debugtoolbar
flask-graphql
requests
pandas
flask-socketio
flask-dropzone
python-magic
flask-compress
gunicorn
sentence-transformers
icecream
google-api-python-client
flasgger
accelerate
